+++ components/atoms/TextInput/text-input.tsx
@@ -1,24 +1,16 @@
 import { CheckCircleFillIcon, XCircleFillIcon } from "@primer/octicons-react";
-import React, { useRef, useState } from "react";
+import React, { useRef } from "react";
 import clsx from "clsx";
-interface TextInputProps {
+
+interface TextInputProps extends React.InputHTMLAttributes<HTMLInputElement> {
   label?: string;
-  placeholder?: string;
-  name?: string;
   state?: "default" | "valid" | "invalid";
-  id?: string;
-  disabled?: boolean;
-  autoFocus?: boolean;
   borderless?: boolean;
   descriptionText?: string;
   classNames?: string;
   errorMsg?: string;
-  value?: string;
-  defaultValue?: string;
-  required?: boolean;
   fieldRef?: React.RefObject<HTMLInputElement>;
-  pattern?: string;
-  onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  handleChange?: (value: string) => void;
 }
 
 const TextInput = ({
@@ -27,34 +19,38 @@ const TextInput = ({
   placeholder,
   state = "default",
   id,
+  value,
   descriptionText,
   classNames,
-  disabled = false,
-  autoFocus,
-  borderless = false,
-  value,
-  defaultValue,
-  onChange,
-  required,
   fieldRef,
-  pattern,
-  errorMsg = ""
+  disabled = false,
+  borderless = false,
+  handleChange,
+  errorMsg = "",
+  ...props
 }: TextInputProps) => {
   const inputRef = useRef<HTMLInputElement>(null);
+
   const handleResetInput = () => {
-    fieldRef ? fieldRef.current!.value = ""
-      :
-      inputRef.current!.value = "";
+    handleChange?.("");
+    if (fieldRef) {
+      fieldRef.current!.value = "";
+    }
   };
+
+  const handleChangeState = (event: React.ChangeEvent<HTMLInputElement>) => {
+    handleChange?.(event.target.value);
+  };
+
   return (
     <>
       <label className="flex w-full flex-col">
-        {label && <p className="mb-2   text-light-slate-9 text-sm">{label}</p>}
+        {label && <p className="mb-2 text-light-slate-9 text-sm">{label}</p>}
         <div
           className={clsx(
             "flex-1 px-3 text-light-slate-12 bg-white shadow-input border transition rounded-lg py-1 flex items-center",
             borderless && "!border-none",
-            state === "invalid" ? " focus-within:border-light-red-10 " : "focus-within:border-light-orange-9 ",
+            state === "invalid" ? "focus-within:border-light-red-10" : "focus-within:border-light-orange-9 ",
             disabled && "bg-light-slate-3 text-light-slate-6",
             classNames
           )}
@@ -65,23 +61,20 @@ const TextInput = ({
             name={name}
             id={id || name || ""}
             placeholder={placeholder || ""}
-            onChange={onChange}
-            value={value}
-            defaultValue={defaultValue}
             className={`flex-1 focus:outline-none  ${classNames} ${
               disabled && "bg-light-slate-3 cursor-not-allowed  text-light-slate-9"
             }`}
-            autoFocus={autoFocus}
             disabled={disabled}
-            required={required}
-            pattern={pattern}
+            value={value}
+            onChange={handleChangeState}
+            {...props}
           />
           {!disabled && (
             <>
               {state === "valid" ? (
                 <CheckCircleFillIcon className="text-light-orange-9" size={14} />
               ) : !!value ? (
-                <span className="flex items-center" onClick={() => handleResetInput()}>
+                <span className="flex items-center" onClick={handleResetInput}>
                   <XCircleFillIcon className="text-light-red-11" size={14} />
                 </span>
               ) : (