+++ components/molecules/HighlightInput/highlight-input-form.tsx
@@ -1,13 +1,22 @@
 import Button from "components/atoms/Button/button";
+import { Textarea } from "components/atoms/Textarea/text-area";
+import { createHighlights } from "lib/hooks/createHighlights";
 
+import { ToastTrigger } from "lib/utils/toast-trigger";
 import { ChangeEvent, useEffect, useRef, useState } from "react";
+import { useSWRConfig } from "swr";
+import useSupabaseAuth from "lib/hooks/useSupabaseAuth";
+import GhOpenGraphImg from "../GhOpenGraphImg/gh-open-graph-img";
 
 const HighlightInputForm = (): JSX.Element => {
+  const { user } = useSupabaseAuth();
+  const { mutate } = useSWRConfig();
   const [isDivFocused, setIsDivFocused] = useState(false);
   const textAreaRef = useRef<HTMLTextAreaElement>(null);
   const [bodyText, setBodyText] = useState("");
   const [row, setRow] = useState(1);
   const [title, setTitle] = useState("");
+  const [pullrequestLink, setPullRequestLink] = useState("");
   const ref = useRef<HTMLFormElement>(null);
   let rowLomit = 5;
   let messageLastScrollHeight = textAreaRef.current ? textAreaRef.current?.scrollHeight : 50;
@@ -21,12 +30,21 @@ const HighlightInputForm = (): JSX.Element => {
       }
     };
     document.addEventListener("mousedown", checkIfClickedOutside);
+    const pullLink = bodyText.match(/((https?:\/\/)?(www\.)?github\.com\/[^\/]+\/[^\/]+\/pull\/[0-9]+)/);
+    const link =
+      pullLink && new URL(pullLink.includes("https://") ? (pullLink as unknown as string) : `https://${pullLink}`);
+
+    if (pullLink && pullLink.length > 0 && link?.hostname === "github.com" && link?.pathname.includes("pull")) {
+      setPullRequestLink(pullLink[0]);
+    } else {
+      setPullRequestLink("");
+    }
 
     return () => {
       // Cleanup the event listener
       document.removeEventListener("mousedown", checkIfClickedOutside);
     };
-  }, [isDivFocused, bodyText]);
+  }, [isDivFocused, bodyText, pullrequestLink]);
 
   const handleTextAreaInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     setBodyText(e.target.value);
@@ -35,22 +53,45 @@ const HighlightInputForm = (): JSX.Element => {
     } else if (row > 1 && textAreaRef.current && textAreaRef.current?.scrollHeight < messageLastScrollHeight) {
       setRow((prev) => prev--);
     }
-
     if (!bodyText) setRow(1);
     messageLastScrollHeight = textAreaRef.current?.scrollHeight || 60;
   };
 
   // Handle submit highlights
-  const handlePostHighlight = (e: React.FormEvent<HTMLFormElement>) => {
+  const handlePostHighlight = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    setIsDivFocused(false);
+
     // Trigger api call to post highlight
-    setBodyText("");
-    setTitle("");
+    const pullLink = bodyText.match(/((https?:\/\/)?(www\.)?github\.com\/[^\/]+\/[^\/]+\/pull\/[0-9]+)/);
+
+    const link = pullLink && new URL(pullLink as unknown as string);
+    const [url] = pullLink || [];
+    const highlight = bodyText.replace(url as string, "");
+
+    if (url === null || url === undefined || url === "" || url.length === 0 || link?.hostname !== "github.com") {
+      ToastTrigger({ message: "A valid Pull request Link is required", type: "error" });
+      return;
+    } else {
+      const res = await createHighlights({
+        highlight,
+        title,
+        url: url || ""
+      });
+
+      if (res) {
+        mutate(`users/${user?.user_metadata.user_name}/highlights`);
+        setBodyText("");
+        setTitle("");
+        setIsDivFocused(false);
+        ToastTrigger({ message: "Highlight uploade success", type: "success" });
+      } else {
+        ToastTrigger({ message: "An error occured!!!", type: "error" });
+      }
+    }
   };
 
   return (
-    <form onSubmit={handlePostHighlight} ref={ref} className="flex flex-col gap-4">
+    <form onSubmit={handlePostHighlight} ref={ref} className="flex flex-1 flex-col gap-4">
       <div
         onClick={() => {
           setIsDivFocused(true);
@@ -64,16 +105,18 @@ const HighlightInputForm = (): JSX.Element => {
           type="text"
           placeholder={isDivFocused ? "Add title (optional)" : "Highlight your merged PRs and provide a link!"}
         />
-        <textarea
-          rows={row}
-          value={bodyText}
-          onChange={(e) => handleTextAreaInputChange(e)}
-          ref={textAreaRef}
+        <Textarea
           className={`resize-none font-normal text-light-slate-11 mb-2 transition focus:outline-none rounded-lg ${
             !isDivFocused ? "hidden" : ""
           }`}
+          ref={textAreaRef}
+          rows={row}
+          value={bodyText}
+          onChange={(e) => handleTextAreaInputChange(e)}
         />
       </div>
+
+      {pullrequestLink && isDivFocused && <GhOpenGraphImg githubLink={pullrequestLink} />}
 
       {isDivFocused && (
         <Button disabled={!bodyText} className="ml-auto" type="primary">