+++ components/organisms/InsightPage/InsightPage.tsx
@@ -37,7 +37,7 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
   }
 
   const [name, setName] = useState(insight?.name || "");
-  const [nameError, setNameError] = useState("");
+  const [isNameValid, setIsNameValid] = useState(false);
   const [submitted, setSubmitted] = useState(false);
   const [repoToAdd, setRepoToAdd] = useState("");
   const [repos, setRepos] = useState<DbRepo[]>(receivedData);
@@ -69,18 +69,27 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
     };
   });
 
-  const handleOnNameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setName(event.target.value);
+  const validateName = (name: string) => {
+    if (!name || name.trim().length <= 3) return false;
+
+    return true;
+  };
+
+  const handleOnNameChange = (value: string) => {
+    setName(value);
+    setIsNameValid(validateName(value));
+  };
+
+  const disableCreateButton = () => {
+    if (insight?.name && validateName(name)) return false;
+    if (submitted) return true;
+    if (!isNameValid) return true;
+
+    return false;
   };
 
   const handleCreateInsightPage = async () => {
     setSubmitted(true);
-
-    if (!name) {
-      setNameError("Insight name is a required field");
-      setSubmitted(false);
-      return;
-    }
 
     const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/user/insights`, {
       method: "POST",
@@ -106,12 +115,6 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
   const handleUpdateInsightPage = async () => {
     setSubmitted(true);
 
-    if (!name) {
-      setNameError("Insight name is a required field");
-      setSubmitted(false);
-      return;
-    }
-
     const response = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/user/insights/${insight?.id}`, {
       method: "PATCH",
       headers: {
@@ -133,9 +136,7 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
     setSubmitted(false);
   };
 
-  const handleOnRepoChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setRepoToAdd(event.target.value);
-  };
+  const handleOnRepoChange = (value: string) => setRepoToAdd(value);
 
   const loadAndAddRepo = async (repoToAdd: string) => {
     setAddRepoError(RepoLookupError.Initial);
@@ -229,8 +230,11 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
             Page Name
           </Title>
 
-          <TextInput placeholder="Page Name (ex: My Team)" value={name} onChange={handleOnNameChange} />
-          {submitted && nameError ? <Text>{nameError}</Text> : ""}
+          <TextInput
+            placeholder="Page Name (ex: My Team)"
+            value={name}
+            handleChange={handleOnNameChange}
+          />
           {/* <Text>insights.opensauced.pizza/pages/{username}/{`{pageId}`}/dashboard</Text> */}
         </div>
 
@@ -242,7 +246,7 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
           <TextInput
             value={repoToAdd}
             placeholder="Repository Full Name (ex: open-sauced/open-sauced)"
-            onChange={handleOnRepoChange}
+            handleChange={handleOnRepoChange}
           />
 
           <div>
@@ -283,7 +287,7 @@ const InsightPage = ({ edit, insight, pageRepos }: InsightPageProps) => {
           handleUpdatePage={handleUpdateInsightPage}
           handleAddToCart={handleReAddRepository}
           history={reposRemoved}
-          createPageButtonDisabled={submitted}
+          createPageButtonDisabled={disableCreateButton()}
         >
           {repos.map((repo) => {
             const totalPrs =