+++ pages/user/settings.tsx
@@ -5,6 +5,7 @@ import ProfileLayout from "layouts/profile";
 import Head from "next/head";
 import { WithPageLayout } from "interfaces/with-page-layout";
 
+
 const ProfileSettings: WithPageLayout = () => {
   const { user } = useSupabaseAuth();
 
@@ -12,7 +13,7 @@ const ProfileSettings: WithPageLayout = () => {
     <>
       {user ? (
         <div className="w-full px-4 md:px-16 lg:px-48 py-16">
-          <UserSettingsPage  user={user} />
+          <UserSettingsPage user={user} />
         </div>
       ) : (
         ""